requirements.txt
lpips==0.1.4
matplotlib==3.10.8
network==0.1
numpy==2.3.5
opencv_python_headless==4.12.0.88
Pillow==12.0.0
pyFFTW==0.15.0
scipy==1.16.3
skimage==0.0
torch==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
